+++ client/consul_discovery.go
@@ -191,9 +191,7 @@ func (d *ConsulDiscovery) watch() {
 					ch := ch
 					go func() {
 						defer func() {
-							if r := recover(); r != nil {
-
-							}
+							recover()
 						}()
 						select {
 						case ch <- pairs: