+++ client/nacos_discovery.go
@@ -174,9 +174,7 @@ func (d *NacosDiscovery) watch() {
 				ch := ch
 				go func() {
 					defer func() {
-						if r := recover(); r != nil {
-
-						}
+						recover()
 					}()
 					select {
 					case ch <- d.pairs: