+++ client/zookeeper_discovery.go
@@ -202,9 +202,7 @@ func (d *ZookeeperDiscovery) watch() {
 					ch := ch
 					go func() {
 						defer func() {
-							if r := recover(); r != nil {
-
-							}
+							recover()
 						}()
 						select {
 						case ch <- pairs: