+++ protocol/message_test.go
@@ -32,7 +32,7 @@ func TestMessage(t *testing.T) {
 	req.Payload = []byte(payload)
 
 	var buf bytes.Buffer
-	err := req.WriteTo(&buf)
+	_, err := req.WriteTo(&buf)
 	if err != nil {
 		t.Fatal(err)
 	}