+++ server/server.go
@@ -156,8 +156,10 @@ func (s *Server) SendMessage(conn net.Conn, servicePath, serviceMethod string, m
 	req.Metadata = metadata
 	req.Payload = data
 
-	reqData := req.Encode()
-	_, err := conn.Write(reqData)
+	b := req.EncodeSlicePointer()
+	_, err := conn.Write(*b)
+	protocol.PutData(b)
+
 	s.Plugins.DoPostWriteRequest(ctx, req, err)
 	protocol.FreeMsg(req)
 	return err
@@ -380,10 +382,10 @@ func (s *Server) serveConn(conn net.Conn) {
 					res.SetCompressType(req.CompressType())
 				}
 				handleError(res, err)
-				data := res.Encode()
-
 				s.Plugins.DoPreWriteResponse(ctx, req, res)
-				conn.Write(data)
+				data := res.EncodeSlicePointer()
+				_, err := conn.Write(*data)
+				protocol.PutData(data)
 				s.Plugins.DoPostWriteResponse(ctx, req, res, err)
 				protocol.FreeMsg(res)
 			} else {
@@ -403,8 +405,9 @@ func (s *Server) serveConn(conn net.Conn) {
 
 			if req.IsHeartbeat() {
 				req.SetMessageType(protocol.Response)
-				data := req.Encode()
-				conn.Write(data)
+				data := req.EncodeSlicePointer()
+				conn.Write(*data)
+				protocol.PutData(data)
 				return
 			}
 
@@ -438,9 +441,9 @@ func (s *Server) serveConn(conn net.Conn) {
 				if len(res.Payload) > 1024 && req.CompressType() != protocol.None {
 					res.SetCompressType(req.CompressType())
 				}
-				data := res.Encode()
-				conn.Write(data)
-				//res.WriteTo(conn)
+				data := res.EncodeSlicePointer()
+				conn.Write(*data)
+				protocol.PutData(data)
 			}
 			s.Plugins.DoPostWriteResponse(newCtx, req, res, err)
 
@@ -526,10 +529,20 @@ func (s *Server) handleRequest(ctx context.Context, req *protocol.Message) (res 
 
 	replyv := argsReplyPools.Get(mtype.ReplyType)
 
+	argv, err = s.Plugins.DoPreCall(ctx, serviceName, methodName, argv)
+	if err != nil {
+		argsReplyPools.Put(mtype.ReplyType, replyv)
+		return handleError(res, err)
+	}
+
 	if mtype.ArgType.Kind() != reflect.Ptr {
 		err = service.call(ctx, mtype, reflect.ValueOf(argv).Elem(), reflect.ValueOf(replyv))
 	} else {
 		err = service.call(ctx, mtype, reflect.ValueOf(argv), reflect.ValueOf(replyv))
+	}
+
+	if err == nil {
+		replyv, err = s.Plugins.DoPostCall(ctx, serviceName, methodName, argv, replyv)
 	}
 
 	argsReplyPools.Put(mtype.ArgType, argv)
@@ -588,7 +601,11 @@ func (s *Server) handleRequestForFunction(ctx context.Context, req *protocol.Mes
 
 	replyv := argsReplyPools.Get(mtype.ReplyType)
 
-	err = service.callForFunction(ctx, mtype, reflect.ValueOf(argv), reflect.ValueOf(replyv))
+	if mtype.ArgType.Kind() != reflect.Ptr {
+		err = service.callForFunction(ctx, mtype, reflect.ValueOf(argv).Elem(), reflect.ValueOf(replyv))
+	} else {
+		err = service.callForFunction(ctx, mtype, reflect.ValueOf(argv), reflect.ValueOf(replyv))
+	}
 
 	argsReplyPools.Put(mtype.ArgType, argv)
 
@@ -698,6 +715,7 @@ func (s *Server) Shutdown(ctx context.Context) error {
 		// wait all in-processing requests finish.
 		ticker := time.NewTicker(shutdownPollInterval)
 		defer ticker.Stop()
+	outer:
 		for {
 			if s.checkProcessMsg() {
 				break
@@ -705,7 +723,7 @@ func (s *Server) Shutdown(ctx context.Context) error {
 			select {
 			case <-ctx.Done():
 				err = ctx.Err()
-				break
+				break outer
 			case <-ticker.C:
 			}
 		}
@@ -736,10 +754,7 @@ func (s *Server) Shutdown(ctx context.Context) error {
 func (s *Server) checkProcessMsg() bool {
 	size := atomic.LoadInt32(&s.handlerMsgNum)
 	log.Info("need handle in-processing msg size:", size)
-	if size == 0 {
-		return true
-	}
-	return false
+	return size == 0
 }
 
 func (s *Server) closeDoneChanLocked() {