+++ serverplugin/mdns.go
@@ -102,10 +102,11 @@ func (p *MDNSRegisterPlugin) Start() error {
 
 // Stop unregister all services.
 func (p *MDNSRegisterPlugin) Stop() error {
+	p.server.Shutdown()
+
 	close(p.dying)
 	<-p.done
 
-	p.server.Shutdown()
 	return nil
 }
 func (p *MDNSRegisterPlugin) initMDNS() {
@@ -146,7 +147,7 @@ func (p *MDNSRegisterPlugin) HandleConnAccept(conn net.Conn) (net.Conn, bool) {
 // Register handles registering event.
 // this service is registered at BASE/serviceName/thisIpAddress node
 func (p *MDNSRegisterPlugin) Register(name string, rcvr interface{}, metadata string) (err error) {
-	if "" == strings.TrimSpace(name) {
+	if strings.TrimSpace(name) == "" {
 		err = errors.New("Register service `name` can't be empty")
 		return
 	}
@@ -175,7 +176,7 @@ func (p *MDNSRegisterPlugin) RegisterFunction(serviceName, fname string, fn inte
 }
 
 func (p *MDNSRegisterPlugin) Unregister(name string) (err error) {
-	if "" == strings.TrimSpace(name) {
+	if strings.TrimSpace(name) == "" {
 		err = errors.New("Register service `name` can't be empty")
 		return
 	}