+++ util/compress.go
@@ -4,15 +4,50 @@ import (
 	"bytes"
 	"compress/gzip"
 	"io/ioutil"
+	"sync"
 )
+
+var (
+	spWriter sync.Pool
+	spReader sync.Pool
+	spBuffer sync.Pool
+)
+
+func init() {
+	spWriter = sync.Pool{New: func() interface{} {
+		return new(gzip.Writer)
+	}}
+	spReader = sync.Pool{New: func() interface{} {
+		return new(gzip.Reader)
+	}}
+	spBuffer = sync.Pool{New: func() interface{} {
+		return bytes.NewBuffer(nil)
+	}}
+}
 
 // Unzip unzips data.
 func Unzip(data []byte) ([]byte, error) {
-	gr, err := gzip.NewReader(bytes.NewBuffer(data))
+	buf := spBuffer.Get().(*bytes.Buffer)
+	defer func() {
+		buf.Reset()
+		spBuffer.Put(buf)
+	}()
+
+	_, err := buf.Write(data)
+	if err != nil {
+		return nil, err
+	}
+
+	gr := spReader.Get().(*gzip.Reader)
+	defer func() {
+		spReader.Put(gr)
+	}()
+	err = gr.Reset(buf)
 	if err != nil {
 		return nil, err
 	}
 	defer gr.Close()
+
 	data, err = ioutil.ReadAll(gr)
 	if err != nil {
 		return nil, err
@@ -22,8 +57,16 @@ func Unzip(data []byte) ([]byte, error) {
 
 // Zip zips data.
 func Zip(data []byte) ([]byte, error) {
-	var buf bytes.Buffer
-	w := gzip.NewWriter(&buf)
+	buf := spBuffer.Get().(*bytes.Buffer)
+	w := spWriter.Get().(*gzip.Writer)
+	w.Reset(buf)
+
+	defer func() {
+		buf.Reset()
+		spBuffer.Put(buf)
+		w.Close()
+		spWriter.Put(w)
+	}()
 	_, err := w.Write(data)
 	if err != nil {
 		return nil, err
@@ -36,6 +79,6 @@ func Zip(data []byte) ([]byte, error) {
 	if err != nil {
 		return nil, err
 	}
-
-	return buf.Bytes(), nil
+	dec := buf.Bytes()
+	return dec, nil
 }